requirements.txt
flask
flask_sqlalchemy
gunicorn
cloudinary
psycopg2-binary
python-dotenv
